+++ ImagePassHandler/CartoonPass/cartoonpass.hh
@@ -37,7 +37,7 @@ namespace bold
       // Do this each frame, as label definitions can change at runtime
       for (std::shared_ptr<PixelLabel> const& label : d_labels)
       {
-        d_bgrByLabelId[label->getID()] = label->modalColour().toBgr();
+        d_bgrByLabelId[(uint8_t)label->getID()] = label->modalColour().toBgr();
       }
       timer.timeEvent("Clear");
     }
