+++ ImagePassHandler/FieldEdgePass/PeriodicFieldEdgePass/periodicfieldedgepass.cc
@@ -9,10 +9,11 @@ using namespace bold;
 using namespace Eigen;
 using namespace std;
 
-PeriodicFieldEdgePass::PeriodicFieldEdgePass(shared_ptr<PixelLabel> fieldLabel, shared_ptr<PixelLabel> lineLabel, ushort pixelWidth, ushort pixelHeight, ushort period)
+PeriodicFieldEdgePass::PeriodicFieldEdgePass(shared_ptr<PixelLabel> fieldLabel, shared_ptr<PixelLabel> lineLabel,
+                                             ushort pixelWidth, ushort pixelHeight, ushort period)
 : FieldEdgePass(pixelWidth, pixelHeight),
-  d_fieldLabelId(fieldLabel->getID()),
-  d_lineLabelId(lineLabel->getID()),
+  d_fieldLabelId((uint8_t)fieldLabel->getID()),
+  d_lineLabelId((uint8_t)lineLabel->getID()),
   d_maxYByC((pixelWidth / period) + 1),
   d_maxYByCConvex((pixelWidth / period) + 1),
   d_runByC((pixelWidth / period) + 1),
